requirements.txt
asgiref==3.2.3
Django==3.0.2
django-environ==0.4.5
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
